moha + chai, need to solve the issue of scss styles in components
