+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "private": false,
   "scripts": {
-    "test": "NODE_ENV=test jasmine JASMINE_CONFIG_PATH='./tests/jasmine.json'",
+    "test": "mocha --opts ./tests/mocha.opts",
     "test:ci:unit": "npm test && npm run test:ci:lint",
     "test:ci:functional": "node ./nightwatch.js -c ./nightwatch.json -e production",
     "test:ci:lint": "eslint ./src/**/*.js",
@@ -96,6 +96,7 @@
     "sass-loader": "2.0.1",
     "serialize-javascript": "^1.0.0",
     "serve-favicon": "^2.2.1",
+    "sinon-chai": "^2.8.0",
     "sitemap": "^1.5.0",
     "strip-loader": "^0.1.2",
     "style-loader": "^0.13.1",
@@ -113,6 +114,7 @@
     "babel-plugin-react-transform": "^2.0.2",
     "babel-preset-react-hmre": "^1.1.1",
     "chai": "^3.0.0",
+    "chai-spies": "^0.7.1",
     "chromedriver": "^2.19.0",
     "del": "^2.0.2",
     "enzyme": "^2.2.0",
@@ -122,8 +124,6 @@
     "eslint-plugin-import": "^1.8.1",
     "eslint-plugin-jsx-a11y": "^1.5.3",
     "eslint-plugin-react": "^5.2.2",
-    "jasmine": "^2.4.1",
-    "jasmine-spec-reporter": "^2.5.0",
     "jscs": "^2.1.1",
     "jsdom": "^8.0.2",
     "mocha": "^2.2.5",
