+++ src/components/Ayah/spec.js
@@ -11,15 +11,15 @@ describe("Ayah", () => {
   });
 
   it('should render', () => {
-    expect(wrapper).toBeDefined();
+    expect(wrapper).to.be.ok;
   });
 
   it('should have correct ayah number', () => {
-    expect(wrapper.find('.label').text()).toBe(ayah.ayahKey);
+    expect(wrapper.find('.label').text()).to.equal(ayah.ayahKey);
   });
 
   it('should contain translations', () => {
-    expect(wrapper.find('.translation').text()).toBe(ayah.content[0].resource.name);
+    expect(wrapper.find('.translation').text()).to.equal(ayah.content[0].resource.name);
   });
