+++ src/components/ContentDropdown/spec.js
@@ -6,7 +6,7 @@ import ContentDropdown, { slugs } from './index';
 
 let wrapper;
 let onOptionChange = sinon.stub();
-let defaultOption = 19
+let defaultOption = 19;
 
 describe('<ContentDropdown />', () => {
   beforeEach(() => {
@@ -15,35 +15,35 @@ describe('<ContentDropdown />', () => {
   });
 
   it('should render', () => {
-    expect(wrapper).toBeDefined();
+    expect(wrapper).to.be.ok;
   });
 
   it('should contain all the content options', () => {
     slugs.filter(slug => slug.language !== 'ar').forEach(slug => {
-      expect(wrapper.text()).toContain(slug.name);
+      expect(wrapper.text()).to.contain(slug.name);
     });
   });
 
   it('should show chosen content option', () => {
-    expect(wrapper.find(`#${defaultOption}en`).prop('checked')).toBe(defaultOption);
+    expect(wrapper.find(`#${defaultOption}en`).prop('checked')).to.equal(defaultOption);
   });
 
-  fit('should add option when clicked', () => {
+  it('should add option when clicked', () => {
     const id = 18;
     wrapper.find(`#${id}en`).simulate('change');
 
-    expect(onOptionChange).toHaveBeenCalled();
-    //expect(onOptionChange).toHaveBeenCalledWith({content: [defaultOption, id]});
+    expect(onOptionChange).to.have.been.called;
+    expect(onOptionChange).to.have.been.calledWith({content: [defaultOption, id]});
   });
 
-  xit('should remove option when clicked', () => {
+  it('should remove option when clicked', () => {
     wrapper.find(`#${defaultOption}en`).simulate('change');
 
     expect(onOptionChange).to.have.been.called;
     expect(onOptionChange).to.have.been.calledWith({content: []});
   });
 
-  xit('should remove all content', () => {
+  it('should remove all content', () => {
     const removeAll = wrapper.find({eventKey: 1})
     expect(removeAll.html()).to.contain('Remove all');
