+++ src/components/FontStyles/index.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { fontFaceStyle, fontFaceStyleLoaded } from '../../helpers/buildFontFaces';
 import { load } from 'redux/modules/fontFaces';
 
-import debug from 'helpers/debug';
+import debug from '../../helpers/debug';
 import selector from './selector';
 
 @connect(
