+++ src/containers/Surah/connect.js
@@ -1,7 +1,7 @@
 import { isAllLoaded, loadAll, setCurrent as setCurrentSurah } from '../../redux/modules/surahs';
 import { clearCurrent, isLoaded, load as loadAyahs } from '../../redux/modules/ayahs';
 
-import debug from 'helpers/debug';
+import debug from '../../helpers/debug';
 
 const ayahRangeSize = 30;
